+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/Camera/HDCamera.cs
@@ -118,6 +118,13 @@ namespace UnityEngine.Rendering.HighDefinition
         // State needed to handle TAAU.
         internal bool previousFrameWasTAAUpsampled = false;
 
+        /// <summary>Ray tracing acceleration structure that is used in case the user specified the build mode as manual for the RTAS.</summary>
+        public RayTracingAccelerationStructure rayTracingAccelerationStructure = null;
+        /// <summary>Flag that tracks if one of the objects that is included into the RTAS had its transform changed.</summary>
+        public bool transformsDirty = false;
+        /// <summary>Flag that tracks if one of the objects that is included into the RTAS had its material changed.</summary>
+        public bool materialsDirty = false;
+
         // Pass all the systems that may want to initialize per-camera data here.
         // That way you will never create an HDCamera and forget to initialize the data.
         /// <summary>