+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/PathTracing/PathTracing.cs
@@ -195,17 +195,17 @@ namespace UnityEngine.Rendering.HighDefinition
             }
 
             // Check materials dirtiness
-            if (GetMaterialDirtiness())
+            if (GetMaterialDirtiness(hdCamera))
             {
-                ResetMaterialDirtiness();
+                ResetMaterialDirtiness(hdCamera);
                 ResetPathTracing();
                 return camData;
             }
 
             // Check light or geometry transforms dirtiness
-            if (GetTransformDirtiness())
+            if (GetTransformDirtiness(hdCamera))
             {
-                ResetTransformDirtiness();
+                ResetTransformDirtiness(hdCamera);
                 ResetPathTracing();
                 return camData;
             }
@@ -219,7 +219,7 @@ namespace UnityEngine.Rendering.HighDefinition
             }
 
             // Check geometry dirtiness
-            ulong accelSize = RequestAccelerationStructure().GetSize();
+            ulong accelSize = RequestAccelerationStructure(hdCamera).GetSize();
             if (accelSize != m_CacheAccelSize)
             {
                 m_CacheAccelSize = accelSize;
@@ -274,7 +274,7 @@ namespace UnityEngine.Rendering.HighDefinition
                 passData.dofParameters = ComputeDoFConstants(hdCamera, m_PathTracingSettings);
                 passData.width = hdCamera.actualWidth;
                 passData.height = hdCamera.actualHeight;
-                passData.accelerationStructure = RequestAccelerationStructure();
+                passData.accelerationStructure = RequestAccelerationStructure(hdCamera);
                 passData.lightCluster = RequestLightCluster();
 
                 passData.shaderVariablesRaytracingCB = m_ShaderVariablesRayTracingCB;